test(ui): Move run creation logic tests out of source files
