+++ ui/src/routes/organizations/$orgId/products/$productId/repositories/$repoId/_repo-layout/create-run/-components/payload.ts
@@ -23,7 +23,6 @@ import {
   PostRepositoryRun,
   ReporterJobConfiguration,
 } from '@/api';
-import { defaultValues } from './default-values';
 import { convertArrayToMap } from './form-primitives';
 import {
   createPluginPayload,
@@ -350,289 +349,4 @@ export function formValuesToPayload(
     jobConfigContext: values.jobConfigContext,
     environmentConfigPath: values.environmentConfigPath,
   };
-}
-
-if (import.meta.vitest) {
-  const { describe, expect, it } = import.meta.vitest;
-
-  function createFormValues(): CreateRunFormValues {
-    return {
-      ...defaultValues(null, [], [], false, [], []),
-      revision: 'main',
-      path: '',
-    };
-  }
-
-  describe('formValuesToPayload', () => {
-    it('omits the environment config when no environment settings are configured', () => {
-      const values = createFormValues();
-      values.jobConfigs.analyzer.infrastructureServices = [
-        {
-          credentialsTypes: ['NETRC_FILE'],
-          name: 'inherited-service',
-          passwordSecretRef: 'password',
-          url: 'https://example.org/repository',
-          usernameSecretRef: 'username',
-        },
-      ];
-
-      const payload = formValuesToPayload(values);
-
-      expect(payload.jobConfigs.analyzer?.environmentConfig).toBeUndefined();
-    });
-
-    it('preserves multiple environment definitions for the same ecosystem', () => {
-      const values = createFormValues();
-      values.jobConfigs.analyzer.environmentDefinitions = {
-        maven: [
-          {
-            service: 'service-a',
-            id: 'repository-a',
-            mirrorOf: '',
-          },
-          {
-            service: 'service-b',
-            id: 'repository-b',
-            mirrorOf: 'central',
-          },
-        ],
-      };
-
-      const payload = formValuesToPayload(values);
-
-      expect(
-        payload.jobConfigs.analyzer?.environmentConfig?.environmentDefinitions
-      ).toEqual(values.jobConfigs.analyzer.environmentDefinitions);
-    });
-
-    it('adds selected package curation providers', () => {
-      const values = createFormValues();
-      values.jobConfigs.analyzer.packageCurationProviders = ['ClearlyDefined'];
-      values.jobConfigs.analyzer.packageCurationProviderConfig = {
-        ClearlyDefined: {
-          options: {
-            serverUrl: 'https://api.clearlydefined.io',
-          },
-          secrets: {
-            token: 'clearly-defined-token',
-          },
-        },
-        File: {
-          options: {
-            path: 'curations.yml',
-          },
-          secrets: {},
-        },
-      };
-
-      const payload = formValuesToPayload(values);
-
-      expect(payload.jobConfigs.analyzer?.packageCurationProviders).toEqual([
-        {
-          type: 'ClearlyDefined',
-          id: 'ClearlyDefined',
-          enabled: true,
-          options: {
-            serverUrl: 'https://api.clearlydefined.io',
-          },
-          secrets: {
-            token: 'clearly-defined-token',
-          },
-        },
-      ]);
-    });
-
-    it('omits package curation providers if none are selected', () => {
-      const values = createFormValues();
-      values.jobConfigs.analyzer.packageCurationProviders = [];
-      values.jobConfigs.analyzer.packageCurationProviderConfig = {
-        ClearlyDefined: {
-          options: {
-            serverUrl: 'https://api.clearlydefined.io',
-          },
-          secrets: {},
-        },
-      };
-
-      const payload = formValuesToPayload(values);
-
-      expect(
-        payload.jobConfigs.analyzer?.packageCurationProviders
-      ).toBeUndefined();
-    });
-
-    it('adds selected evaluator package configuration providers', () => {
-      const values = createFormValues();
-      values.jobConfigs.evaluator.packageConfigurationProviders = ['OrtConfig'];
-      values.jobConfigs.evaluator.packageConfigurationProviderConfig = {
-        OrtConfig: {
-          options: {
-            path: 'evaluator-package-configurations.yml',
-          },
-          secrets: {
-            token: 'evaluator-token',
-          },
-        },
-        File: {
-          options: {
-            path: 'ignored-package-configurations.yml',
-          },
-          secrets: {},
-        },
-      };
-
-      const payload = formValuesToPayload(values);
-
-      expect(
-        payload.jobConfigs.evaluator?.packageConfigurationProviders
-      ).toEqual([
-        {
-          type: 'OrtConfig',
-          id: 'OrtConfig',
-          enabled: true,
-          options: {
-            path: 'evaluator-package-configurations.yml',
-          },
-          secrets: {
-            token: 'evaluator-token',
-          },
-        },
-      ]);
-    });
-
-    it('adds selected reporter package configuration providers when evaluator is disabled', () => {
-      const values = createFormValues();
-      values.jobConfigs.evaluator.enabled = false;
-      values.jobConfigs.reporter.packageConfigurationProviders = ['OrtConfig'];
-      values.jobConfigs.reporter.packageConfigurationProviderConfig = {
-        OrtConfig: {
-          options: {
-            path: 'reporter-package-configurations.yml',
-          },
-          secrets: {
-            token: 'reporter-token',
-          },
-        },
-      };
-
-      const payload = formValuesToPayload(values);
-
-      expect(payload.jobConfigs.evaluator).toBeUndefined();
-      expect(
-        payload.jobConfigs.reporter?.packageConfigurationProviders
-      ).toEqual([
-        {
-          type: 'OrtConfig',
-          id: 'OrtConfig',
-          enabled: true,
-          options: {
-            path: 'reporter-package-configurations.yml',
-          },
-          secrets: {
-            token: 'reporter-token',
-          },
-        },
-      ]);
-    });
-
-    it('omits package configuration providers if none are selected', () => {
-      const values = createFormValues();
-      values.jobConfigs.evaluator.packageConfigurationProviders = [];
-      values.jobConfigs.evaluator.packageConfigurationProviderConfig = {
-        OrtConfig: {
-          options: {
-            path: 'evaluator-package-configurations.yml',
-          },
-          secrets: {},
-        },
-      };
-      values.jobConfigs.reporter.packageConfigurationProviders = [];
-      values.jobConfigs.reporter.packageConfigurationProviderConfig = {
-        OrtConfig: {
-          options: {
-            path: 'reporter-package-configurations.yml',
-          },
-          secrets: {},
-        },
-      };
-
-      const payload = formValuesToPayload(values);
-
-      expect(
-        payload.jobConfigs.evaluator?.packageConfigurationProviders
-      ).toBeUndefined();
-      expect(
-        payload.jobConfigs.reporter?.packageConfigurationProviders
-      ).toBeUndefined();
-    });
-
-    it('omits reporter package configuration providers when evaluator is enabled', () => {
-      const values = createFormValues();
-      values.jobConfigs.evaluator.packageConfigurationProviders = ['OrtConfig'];
-      values.jobConfigs.evaluator.packageConfigurationProviderConfig = {
-        OrtConfig: {
-          options: {
-            path: 'evaluator-package-configurations.yml',
-          },
-          secrets: {},
-        },
-      };
-      values.jobConfigs.reporter.packageConfigurationProviders = ['File'];
-      values.jobConfigs.reporter.packageConfigurationProviderConfig = {
-        File: {
-          options: {
-            path: 'reporter-package-configurations.yml',
-          },
-          secrets: {},
-        },
-      };
-
-      const payload = formValuesToPayload(values);
-
-      expect(
-        payload.jobConfigs.evaluator?.packageConfigurationProviders
-      ).toEqual([
-        {
-          type: 'OrtConfig',
-          id: 'OrtConfig',
-          enabled: true,
-          options: {
-            path: 'evaluator-package-configurations.yml',
-          },
-        },
-      ]);
-      expect(
-        payload.jobConfigs.reporter?.packageConfigurationProviders
-      ).toBeUndefined();
-    });
-
-    it('omits infrastructure services from environment configs', () => {
-      const values = createFormValues();
-      values.jobConfigs.analyzer.environmentDefinitions = {
-        conan: [
-          {
-            service: 'service-a',
-            name: 'private-conan',
-            url: '',
-            verifySsl: 'true',
-          },
-        ],
-      };
-      values.jobConfigs.analyzer.infrastructureServices = [
-        {
-          credentialsTypes: ['NETRC_FILE'],
-          name: 'inherited-service',
-          passwordSecretRef: 'password',
-          url: 'https://example.org/repository',
-          usernameSecretRef: 'username',
-        },
-      ];
-
-      const payload = formValuesToPayload(values);
-
-      expect(payload.jobConfigs.analyzer?.environmentConfig).not.toHaveProperty(
-        'infrastructureServices'
-      );
-    });
-  });
 }
