+++ ui/src/routes/organizations/$orgId/products/$productId/repositories/$repoId/_repo-layout/create-run/-components/plugin-utils.ts
@@ -389,151 +389,3 @@ export function createProviderPluginPayload(
     };
   });
 }
-
-if (import.meta.vitest) {
-  const { expect, it } = import.meta.vitest;
-
-  it('getPluginDefaultValues ignores default values for secret options', () => {
-    const defaults = getPluginDefaultValues([
-      {
-        id: 'SCANOSS',
-        type: 'SCANNER',
-        displayName: 'SCANOSS',
-        summary: 'A scanner plugin.',
-        description: 'A scanner plugin.',
-        options: [
-          {
-            name: 'apiKey',
-            description: 'The API key.',
-            type: 'SECRET',
-            defaultValue: 'server-side-default-secret',
-            isFixed: false,
-            isNullable: false,
-            isRequired: false,
-          },
-        ],
-      },
-    ]);
-
-    expect(defaults.SCANOSS?.secrets).toEqual({});
-  });
-
-  it('providerPluginConfigsToFormValues reconstructs selected providers and config', () => {
-    const formValues = providerPluginConfigsToFormValues([
-      {
-        type: 'ClearlyDefined',
-        id: 'ClearlyDefined',
-        enabled: true,
-        options: {
-          serverUrl: 'https://api.clearlydefined.io',
-        },
-        secrets: {
-          token: 'clearly-defined-token',
-        },
-      },
-      {
-        type: 'File',
-        id: 'File',
-        enabled: false,
-      },
-    ]);
-
-    expect(formValues).toEqual({
-      selectedPluginIds: ['ClearlyDefined'],
-      config: {
-        ClearlyDefined: {
-          options: {
-            serverUrl: 'https://api.clearlydefined.io',
-          },
-          secrets: {
-            token: 'clearly-defined-token',
-          },
-        },
-        File: {
-          options: {},
-          secrets: {},
-        },
-      },
-    });
-  });
-
-  it('createPluginPayload omits blank secret values from the payload', () => {
-    const payload = createPluginPayload(
-      {
-        SCANOSS: {
-          options: {
-            apiUrl: 'https://api.osskb.org',
-          },
-          secrets: {
-            apiKey: '',
-          },
-        },
-      },
-      ['SCANOSS']
-    );
-
-    expect(payload).toEqual({
-      SCANOSS: {
-        options: {
-          apiUrl: 'https://api.osskb.org',
-        },
-        secrets: {},
-      },
-    });
-  });
-
-  it('createProviderPluginPayload creates provider configurations for selected plugins', () => {
-    const payload = createProviderPluginPayload(
-      {
-        ClearlyDefined: {
-          options: {
-            serverUrl: 'https://api.clearlydefined.io',
-          },
-          secrets: {
-            token: '',
-          },
-        },
-        File: {
-          options: {
-            path: 'curations.yml',
-          },
-          secrets: {},
-        },
-      },
-      ['ClearlyDefined']
-    );
-
-    expect(payload).toEqual([
-      {
-        type: 'ClearlyDefined',
-        id: 'ClearlyDefined',
-        enabled: true,
-        options: {
-          serverUrl: 'https://api.clearlydefined.io',
-        },
-      },
-    ]);
-  });
-
-  it('reconstructScannerSelection rebuilds scanner scopes for rerun values', () => {
-    const defaults = {
-      scanners: ['ScanCode'],
-      scannerScopes: {
-        ScanCode: 'both' as const,
-      },
-    };
-
-    const result = reconstructScannerSelection(
-      ['SCANOSS'],
-      ['SCANOSS'],
-      defaults
-    );
-
-    expect(result).toEqual({
-      scanners: ['SCANOSS'],
-      scannerScopes: {
-        SCANOSS: 'both',
-      },
-    });
-  });
-}
