+++ ui/tests/unit/logic/plugin-utils.test.ts
@@ -1,0 +1,172 @@
+/*
+ * Copyright (C) 2024 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+import { expect, it } from 'vitest';
+
+import {
+  createPluginPayload,
+  createProviderPluginPayload,
+  getPluginDefaultValues,
+  providerPluginConfigsToFormValues,
+  reconstructScannerSelection,
+} from '@/routes/organizations/$orgId/products/$productId/repositories/$repoId/_repo-layout/create-run/-components/plugin-utils';
+
+it('createPluginPayload omits blank secret values from the payload', () => {
+  const payload = createPluginPayload(
+    {
+      SCANOSS: {
+        options: {
+          apiUrl: 'https://api.osskb.org',
+        },
+        secrets: {
+          apiKey: '',
+        },
+      },
+    },
+    ['SCANOSS']
+  );
+
+  expect(payload).toEqual({
+    SCANOSS: {
+      options: {
+        apiUrl: 'https://api.osskb.org',
+      },
+      secrets: {},
+    },
+  });
+});
+
+it('createProviderPluginPayload creates provider configurations for selected plugins', () => {
+  const payload = createProviderPluginPayload(
+    {
+      ClearlyDefined: {
+        options: {
+          serverUrl: 'https://api.clearlydefined.io',
+        },
+        secrets: {
+          token: '',
+        },
+      },
+      File: {
+        options: {
+          path: 'curations.yml',
+        },
+        secrets: {},
+      },
+    },
+    ['ClearlyDefined']
+  );
+
+  expect(payload).toEqual([
+    {
+      type: 'ClearlyDefined',
+      id: 'ClearlyDefined',
+      enabled: true,
+      options: {
+        serverUrl: 'https://api.clearlydefined.io',
+      },
+    },
+  ]);
+});
+
+it('getPluginDefaultValues ignores default values for secret options', () => {
+  const defaults = getPluginDefaultValues([
+    {
+      id: 'SCANOSS',
+      type: 'SCANNER',
+      displayName: 'SCANOSS',
+      summary: 'A scanner plugin.',
+      description: 'A scanner plugin.',
+      options: [
+        {
+          name: 'apiKey',
+          description: 'The API key.',
+          type: 'SECRET',
+          defaultValue: 'server-side-default-secret',
+          isFixed: false,
+          isNullable: false,
+          isRequired: false,
+        },
+      ],
+    },
+  ]);
+
+  expect(defaults.SCANOSS?.secrets).toEqual({});
+});
+
+it('providerPluginConfigsToFormValues reconstructs selected providers and config', () => {
+  const formValues = providerPluginConfigsToFormValues([
+    {
+      type: 'ClearlyDefined',
+      id: 'ClearlyDefined',
+      enabled: true,
+      options: {
+        serverUrl: 'https://api.clearlydefined.io',
+      },
+      secrets: {
+        token: 'clearly-defined-token',
+      },
+    },
+    {
+      type: 'File',
+      id: 'File',
+      enabled: false,
+    },
+  ]);
+
+  expect(formValues).toEqual({
+    selectedPluginIds: ['ClearlyDefined'],
+    config: {
+      ClearlyDefined: {
+        options: {
+          serverUrl: 'https://api.clearlydefined.io',
+        },
+        secrets: {
+          token: 'clearly-defined-token',
+        },
+      },
+      File: {
+        options: {},
+        secrets: {},
+      },
+    },
+  });
+});
+
+it('reconstructScannerSelection rebuilds scanner scopes for rerun values', () => {
+  const defaults = {
+    scanners: ['ScanCode'],
+    scannerScopes: {
+      ScanCode: 'both' as const,
+    },
+  };
+
+  const result = reconstructScannerSelection(
+    ['SCANOSS'],
+    ['SCANOSS'],
+    defaults
+  );
+
+  expect(result).toEqual({
+    scanners: ['SCANOSS'],
+    scannerScopes: {
+      SCANOSS: 'both',
+    },
+  });
+});
